Crate cargo_edit

Show and Edit Cargo's Manifest Files

Structs

Dependency
A dependency handled by Cargo

Error
The Error type.

Manifest
A Cargo Manifest

Enums

ErrorKind
The kind of an error.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Functions

get_crate_name_from_github
Query crate name by accessing a github repo Cargo.toml

get_crate_name_from_gitlab
Query crate name by accessing a gitlab repo Cargo.toml

get_crate_name_from_path
Query crate name by accessing Cargo.toml in a local path

get_latest_dependency
Query latest version from crates.io

Type Definitions

Result
Convenient wrapper around std::Result.