Function cargo_edit::get_latest_dependency

pub fn get_latest_dependency(
    crate_name: &str, 
    flag_allow_prerelease: bool
) -> Result<Dependency>

Query latest version from crates.io

The latest version will be returned as a Dependency. This will fail, when

there is no Internet connection,

the response from crates.io is an error or in an incorrect format,

or when a crate with the given name does not exist on crates.io.