Struct cargo_edit::LocalManifest

pub struct LocalManifest { /* fields omitted */ }

A Cargo manifest that is available locally.

Methods

impl LocalManifest

pub fn find(path: &Option<PathBuf>) -> Result<Self>

Construct a LocalManifest. If no path is provided, make an educated guess as to which one the user means.

pub fn try_new(path: &Path) -> Result<Self>

Construct the LocalManifest corresponding to the Path provided.

pub fn upgrade(&mut self, dependency: &Dependency, dry_run: bool) -> Result<()>

Instruct this manifest to upgrade a single dependency. If this manifest does not have that dependency, it does nothing.