Struct cargo_edit::Manifest

pub struct Manifest {
    pub data: Table,
}

A Cargo Manifest

Fields

data

Manifest contents as TOML data

Methods

impl Manifest

fn find_file(path: &Option<&str>) -> Result<File, Box<Error>>

Look for a Cargo.toml file

Starts at the given path an goes into its parent directories until the manifest file is found. If no path is given, the process's working directory is used as a starting point.

fn find_lock_file(path: &Option<&str>) -> Result<File, Box<Error>>

Look for a Cargo.lock file

Starts at the given path an goes into its parent directories until the manifest file is found. If no path is given, the process' working directory is used as a starting point.

fn open(path: &Option<&str>) -> Result<Manifest, Box<Error>>

Open the Cargo.toml for a path (or the process' cwd)

fn open_lock_file(path: &Option<&str>) -> Result<Manifest, Box<Error>>

Open the Cargo.lock for a path (or the process' cwd)

fn write_to_file(&self, file: &mut File) -> Result<(), Box<Error>>

Overwrite a file with TOML data.

fn insert_into_table(&mut self, table: &str, dep: &Dependency) -> Result<(), ManifestError>

Add entry to a Cargo.toml.

fn remove_from_table(&mut self, table: &str, name: &str) -> Result<(), ManifestError>

Remove entry from a Cargo.toml.

Examples

    use cargo_edit::{Dependency, Manifest};
    use toml;

    let mut manifest = Manifest { data: toml::Table::new() };
    let dep = Dependency::new("cargo-edit").set_version("0.1.0");
    let _ = manifest.insert_into_table("dependencies", &dep);
    assert!(manifest.remove_from_table("dependencies", &dep.name).is_ok());
    assert!(manifest.remove_from_table("dependencies", &dep.name).is_err());
    assert!(manifest.data.is_empty());

fn add_deps(&mut self, table: &str, deps: &[Dependency]) -> Result<(), Box<Error>>

Add multiple dependencies to manifest